Subscription process modified
Subscribed calendars are created similarly to newly created owned
calendars and they use calEeeManager to immediately update their
properties.  These properties are set in subscription delegate on
calICalendar instance.  They are computed from updated selection
structure which contains owner and calendar data.

+++ src/calEeeManager.js
@@ -177,11 +177,19 @@ calEeeManager.prototype = {
       }
     };
 
-    cal3eRequest.Client.getInstance().deleteCalendar(
-      this._getIdentity(calendar),
-      listener,
-      calendar
-    );
+    if (cal3eModel.isOwnedCalendar(calendar)) {
+      cal3eRequest.Client.getInstance().deleteCalendar(
+        this._getIdentity(calendar),
+        listener,
+        calendar
+      );
+    } else {
+      cal3eRequest.Client.getInstance().unsubscribeCalendar(
+        this._getIdentity(calendar),
+        listener,
+        calendar
+      );
+    }
   },
 
   onPropertyChanged:
